refactor(server): Remove procedure kit compatibility
