+++ integration/next-app/README.md
@@ -38,7 +38,7 @@ Use these files as entry points, depending on what you want to understand:
 - Procedure form-data input via `.handle(...).nextRoute(...)` sugar: `app/api/procedure-form-data/route.ts`
 - Procedure runtime output validation: `app/api/procedure-invalid-output/route.ts`
 - Procedure validator-stage customization: `app/api/procedure-validation-branch/route.ts`
-- Project-level procedure defaults via `procedure.defaults({ onError })`: `app/api/procedure-kit-error/route.ts`
+- Project-level procedure defaults via `procedure.defaults({ onError })`: `app/api/procedure-defaults-error/route.ts`
 - Procedure-first walkthrough page: `app/procedure-examples/page.tsx`
 - Typed client setup: `src/lib/rpc-client.ts`
 - Generated output shape: `src/generated/rpc.ts`
@@ -116,7 +116,7 @@ This workspace is intended to make scanner and runtime regressions visible in Gi
 
 The main walkthrough in this fixture is now procedure-first. `app/api/procedure-contract/[userId]/route.ts` is the baseline typed route: it binds the generated `routeContract`, declares params/query/output in one builder, and exports `GET` through terminal `.handle(...).nextRoute({ onError })` sugar. `app/api/procedure-submit/route.ts` extends that path to json/header/cookie input, `app/api/procedure-form-data/route.ts` covers multipart-style input with the same terminal shape, and `app/api/procedure-guarded/[userId]/route.ts` shows the shared-preset case where keeping a standalone `nextRoute(procedure, options)` call remains natural because the procedure value comes from a reusable base.
 
-The procedure fixtures also cover the later design phases that made the procedure path complete enough to recommend by default. `app/api/procedure-invalid-output/route.ts` demonstrates opt-in runtime output enforcement with a Standard Schema output contract. `app/api/procedure-kit-error/route.ts` shows project-level `procedure.defaults({ onError })` usage, while `app/api/_shared/procedure-kit.ts` keeps `createProcedureKit(...)` as compatibility sugar over the same shared preset model. `app/api/procedure-validation-branch/route.ts` shows validator-stage customization through `procedure.query(schema, { onValidationError(...) { ... } })`. `app/api/error-demo/route.ts` keeps the direct standalone `nextRoute(..., { onError })` form for arbitrary thrown errors, and `app/api/_shared/on-error.ts` shows `rpcError(...)` plus generic `Error` mapping in a shared `onError` implementation.
+The procedure fixtures also cover the later design phases that made the procedure path complete enough to recommend by default. `app/api/procedure-invalid-output/route.ts` demonstrates opt-in runtime output enforcement with a Standard Schema output contract. `app/api/procedure-defaults-error/route.ts` shows project-level `procedure.defaults({ onError })` usage, while `app/api/_shared/procedure-defaults.ts` keeps that shared preset explicit. `app/api/procedure-validation-branch/route.ts` shows validator-stage customization through `procedure.query(schema, { onValidationError(...) { ... } })`. `app/api/error-demo/route.ts` keeps the direct standalone `nextRoute(..., { onError })` form for arbitrary thrown errors, and `app/api/_shared/on-error.ts` shows `rpcError(...)` plus generic `Error` mapping in a shared `onError` implementation.
 
 When a shared handler should contribute its concrete error envelope to client-side
 response inference, define it with `satisfies ProcedureOnError` instead of
