+++ integration/next-app/app/api/procedure-defaults-error/route-contract.ts
@@ -1,9 +1,9 @@
 import type { ProcedureRouteContract } from "rpc4next/server";
 
 export type Params = {};
-export type RouteContract = ProcedureRouteContract<"/api/procedure-kit-error", Params>;
+export type RouteContract = ProcedureRouteContract<"/api/procedure-defaults-error", Params>;
 
 export const routeContract = {
-  pathname: "/api/procedure-kit-error",
+  pathname: "/api/procedure-defaults-error",
   params: {} as Params,
 } as RouteContract;
