+++ packages/rpc4next/src/rpc/server/meta.ts
@@ -1,12 +1,8 @@
 import type {
-  MergeProcedureDefinition,
   ProcedureDefinition,
   WithProcedureDefinition,
 } from "./procedure-types";
-import {
-  attachProcedureDefinition,
-  getProcedureDefinition,
-} from "./procedure-types";
+import { getProcedureDefinition } from "./procedure-types";
 
 export interface RpcMetaBase {
   summary?: string;
@@ -31,27 +27,6 @@ export type InferRouteMeta<TValue> =
   >
     ? TMeta
     : never;
-
-type ReplaceProcedureMeta<
-  TValue,
-  TMeta extends RpcMeta,
-  TDefinition extends ProcedureDefinition = ExtractProcedureDefinition<TValue>,
-> = WithProcedureDefinition<
-  TValue,
-  MergeProcedureDefinition<TDefinition, { meta: TMeta }>
->;
-
-export const withMeta = <TMeta extends RpcMeta, TValue extends object = object>(
-  meta: TMeta,
-  value: TValue,
-): ReplaceProcedureMeta<TValue, TMeta> => {
-  const definition = getProcedureDefinition(value);
-
-  return attachProcedureDefinition(value, {
-    ...definition,
-    meta,
-  }) as ReplaceProcedureMeta<TValue, TMeta>;
-};
 
 export const getRouteMeta = <TValue>(value: TValue): InferRouteMeta<TValue> => {
   return getProcedureDefinition(value)?.meta as InferRouteMeta<TValue>;
